requirements.txt
psycopg2
pendulum
pandas
python-pptx